% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_function.R
\name{catch_function}
\alias{catch_function}
\title{Catch Function}
\usage{
catch_function(scenario, Arrowtooth, Atka, Flathead, Greenland, Kamchatka,
  Northern, Octopus, OtherFlat, OtherRock, PCod, Plaice, POP, Pollock, Rock,
  Rougheye, Sablefish, Sculpin, Shark, Shortraker, Skate, Squid, Yellowfin)
}
\arguments{
\item{scenario}{The economic scenario number. 1: Status Quo (log-linear). 1.1: Status quo (log-log.  This fits just as well as log-linear in retrospect, but assumes if the target dissapears (e.g. Atka) then the catch of the bycatch (e.g. Northern) would go to 0.  This assumption is problematic so this is an "alternative" status quo but not the "go to").  2. Whitefish dominate--the council moves TAC from flatfish to pollock and pcod. 3. Flatfish dominate--the council moves TAC from pollock and pcod to flatfish.  Also assumes 30% increase in flatfish catch over the baseline fishing scenario, or else why would the council move TAC? 4. No fishing (will only return 0's)}

\item{Arrowtooth}{Optional.  ABC of Arrowtooth Flounder.}

\item{Atka}{Optional.  ABC of Atka Mackerel.}

\item{Flathead}{Optional.  ABC of Flathead Sole.}

\item{Greenland}{Optional.  ABC of Greenland Turbot.}

\item{Kamchatka}{Optional.  ABC of Kamchatka Flounder.}

\item{Northern}{Optional.  ABC of Northern Rockfish.}

\item{Octopus}{Optional.  ABC of Octopus.}

\item{OtherFlat}{Optional.  ABC of Other Flatfish.}

\item{OtherRock}{Optional.  ABC of Other Rockfish.}

\item{PCod}{Optional.  ABC of Pacific Cod.}

\item{Plaice}{Optional.  ABC of Alaska Plaice.}

\item{POP}{Optional.  ABC of Pacific Ocean Perch.}

\item{Pollock}{Optional.  ABC of Pollock.}

\item{Rock}{Optional.  ABC of Rock Sole.}

\item{Rougheye}{Optional.  ABC of Rougheye Rockfish.}

\item{Sablefish}{Optional.  ABC of Sablefish.}

\item{Sculpin}{Optional.  ABC of Sculpin.}

\item{Shark}{Optional.  ABC of Shark.}

\item{Shortraker}{Optional.  ABC of Shortraker Rockfish.}

\item{Skate}{Optional.  ABC of Skate.}

\item{Squid}{Optional.  ABC of Squid.}

\item{Yellowfin}{Optional.  ABC of Yellowfin Sole.}
}
\description{
This function predicts the BSAI catch for each species whose ABC is given.  It is meant to work with the ACLIM bio models.

If you have any questions, please contact me (e-mail: amanda.faig@noaa.gov, call: X-4281).

Currently programmed scenarios: Scenario 1, Scenario 1.1 & Scenario 3
}
\examples{
catch_function(1, Pollock = 2e6, Arrowtooth = 2e5, Yellowfin = 2e5)
catch_function(3, Pollock = 2e6, Yellowfin = 2e5, PCod = 1e5)
}
