% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_function.R
\name{catch_function}
\alias{catch_function}
\title{Catch Function}
\usage{
catch_function(scenario, Arrowtooth, Atka, Flathead, Greenland, Kamchatka,
  Northern, Octopus, OtherFlat, OtherRock, PCod, Plaice, POP, Pollock, Rock,
  Rougheye, Sablefish, Sculpin, Shark, Shortraker, Skate, Squid, Yellowfin,
  spptomult, multiplier)
}
\arguments{
\item{scenario}{The economic scenario number. Current options: 1, 1.1, 2, 3, or 4}

\item{Arrowtooth}{Optional.  ABC of Arrowtooth Flounder.}

\item{Atka}{Optional.  ABC of Atka Mackerel.}

\item{Flathead}{Optional.  ABC of Flathead Sole.}

\item{Greenland}{Optional.  ABC of Greenland Turbot.}

\item{Kamchatka}{Optional.  ABC of Kamchatka Flounder.}

\item{Northern}{Optional.  ABC of Northern Rockfish.}

\item{Octopus}{Optional.  ABC of Octopus.}

\item{OtherFlat}{Optional.  ABC of Other Flatfish.}

\item{OtherRock}{Optional.  ABC of Other Rockfish.}

\item{PCod}{Optional.  ABC of Pacific Cod.}

\item{Plaice}{Optional.  ABC of Alaska Plaice.}

\item{POP}{Optional.  ABC of Pacific Ocean Perch.}

\item{Pollock}{Optional.  ABC of Pollock.}

\item{Rock}{Optional.  ABC of Rock Sole.}

\item{Rougheye}{Optional.  ABC of Rougheye Rockfish.}

\item{Sablefish}{Optional.  ABC of Sablefish.}

\item{Sculpin}{Optional.  ABC of Sculpin.}

\item{Shark}{Optional.  ABC of Shark.}

\item{Shortraker}{Optional.  ABC of Shortraker Rockfish.}

\item{Skate}{Optional.  ABC of Skate.}

\item{Squid}{Optional.  ABC of Squid.}

\item{Yellowfin}{Optional.  ABC of Yellowfin Sole.}

\item{spptomult}{Required if running scenario 5.  Will be discarded otherwise.  Choose a species catch to override with N*ABC.  Must be spelt exactly as one of the species parameters, case sensitive.  Must be in quotation marks.  See examples.}

\item{multiplier}{Required if running scenario 5.  Will be discarded otherwise.  The N which will be multiplied with ABC to override the species designated by spptomult.}
}
\description{
This function predicts the BSAI catch for each species whose ABC is given.  It is meant to work with the ACLIM bio models.

If you have any questions, please contact Amanda Faig (e-mail: amanda.faig@noaa.gov, call: X-4281).

This version last updated June 2018

Currently programmed scenarios: \cr
Scenario 1: Status Quo (Log-Linear) \cr
Scenario 1.1: Status Quo alternative: Log-Log \cr
Scenario 2: Whitefish (Pollock and Cod) Political (aka TAC-setting) Preference \cr
Scenario 3: Flatfish Political (aka TAC-setting) Preference \cr
Scenario 4: No Fishing (will return all zeros) \cr
Scenario 5: Fiddle with a single species. \cr
}
\examples{
catch_function(1, Pollock = 2e6, Arrowtooth = 2e5, Yellowfin = 2e5)
catch_function(3, Pollock = 2e6, Yellowfin = 2e5, PCod = 1e5)
catch_function(5, spptomult = "Arrowtooth", multiplier = 2, Pollock = 2e6, Arrowtooth = 2e5, Yellowfin = 2e5)
}
