% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAC_function.R
\name{TAC_function}
\alias{TAC_function}
\title{TAC Function}
\usage{
TAC_function(
  scenario,
  Arrowtooth,
  Atka,
  Flathead,
  Greenland,
  Kamchatka,
  Northern,
  Octopus,
  OtherFlat,
  OtherRock,
  PCod,
  Plaice,
  POP,
  Pollock,
  Rock,
  Rougheye,
  Sablefish,
  Sculpin,
  Shark,
  Shortraker,
  Skate,
  Squid,
  Yellowfin,
  spptomult,
  multiplier
)
}
\arguments{
\item{scenario}{The economic scenario number. Current options: 1, 2, 3, 4, 5.1, 5.2, or 5.3}

\item{Arrowtooth}{Optional.  ABC of Arrowtooth Flounder.}

\item{Atka}{Optional.  ABC of Atka Mackerel.}

\item{Flathead}{Optional.  ABC of Flathead Sole.}

\item{Greenland}{Optional.  ABC of Greenland Turbot.}

\item{Kamchatka}{Optional.  ABC of Kamchatka Flounder.}

\item{Northern}{Optional.  ABC of Northern Rockfish.}

\item{Octopus}{Optional.  ABC of Octopus.}

\item{OtherFlat}{Optional.  ABC of Other Flatfish.}

\item{OtherRock}{Optional.  ABC of Other Rockfish.}

\item{PCod}{Optional.  ABC of Pacific Cod.}

\item{Plaice}{Optional.  ABC of Alaska Plaice.}

\item{POP}{Optional.  ABC of Pacific Ocean Perch.}

\item{Pollock}{Optional.  ABC of Pollock.}

\item{Rock}{Optional.  ABC of Rock Sole.}

\item{Rougheye}{Optional.  ABC of Rougheye Rockfish.}

\item{Sablefish}{Optional.  ABC of Sablefish.}

\item{Sculpin}{Optional.  ABC of Sculpin.}

\item{Shark}{Optional.  ABC of Shark.}

\item{Shortraker}{Optional.  ABC of Shortraker Rockfish.}

\item{Skate}{Optional.  ABC of Skate.}

\item{Squid}{Optional.  ABC of Squid.}

\item{Yellowfin}{Optional.  ABC of Yellowfin Sole.}

\item{spptomult}{Required if running any of the 5-series scenarios.  Will be discarded otherwise.  Choose a species catch to override with N*ABC.  Must be spelt exactly as one of the species parameters, case sensitive.  Must be in quotation marks.  If you want to replace more than one species, create a vector of strings (e.g. c("Arrowtooth","Atka"))}

\item{multiplier}{Required if running scenario 5-series scenarios.  Will be discarded otherwise.  The N which will be multiplied with ABC to override the species designated by spptomult. If you are replacing more than one species, the order of the numbers corresponds to the order of the names in the spptomult string. (e.g. c(1,5) would imply the first species listed in spptomult has its catch replaced with 1*ABC_spp1 and the second is replaced with 5*ABC_spp2)}
}
\description{
This function predicts the BS TAC for each species whose ABC is given.  

If you have any questions, please contact Amanda Faig (e-mail: amanda.faig@noaa.gov, call: X-4281).

This version last updated Feb 2018

Currently programmed scenarios: \cr
Scenario 1: Status Quo (Log-Linear) \cr
Scenario 2: Whitefish (Pollock and Cod) Political (aka TAC-setting) Preference \cr
Scenario 3: Flatfish Political (aka TAC-setting) Preference \cr
Scenario 4: No Fishing (will return all zeros) \cr
Scenario 5.1: Fiddle with a single species--calculate the rest still taking the ABC of the removed sp. in to account. \cr
Scenario 5.2: Fiddle with a single species--calculate the rest assuming the ABC of the removed sp. does not influence the sp. under the cap at all. \cr
Scenario 5.3: Fiddle with a single species--calculate the rest assuming the ABC of the removed sp. does not influence the sp. under the cap at all and then increase the TAC of all the remaining species until the sum of the tAC = 2mmt \cr
}
\examples{
TAC_function(1, Pollock = 2e6, Arrowtooth = 2e5, Yellowfin = 2e5)
TAC_function(3, Pollock = 2e6, Yellowfin = 2e5, PCod = 1e5)
TAC_function(5.1, spptomult = "Arrowtooth", multiplier = 2, 
                 Pollock = 2e6, Arrowtooth = 2e5, Yellowfin = 2e5)
TAC_function(5.1, spptomult = c("Arrowtooth","Yellowfin"), multiplier = c(2,1), 
                 Pollock = 2e6, Arrowtooth = 2e5, Yellowfin = 2e5)
TAC_function(5.2, spptomult = "Arrowtooth", multiplier = 2, 
                 Pollock = 2e6, Arrowtooth = 2e5, Yellowfin = 2e5)
TAC_function(5.2, spptomult = c("Arrowtooth","Yellowfin"), multiplier = c(2,1), 
                 Pollock = 2e6, Arrowtooth = 2e5, Yellowfin = 2e5)
TAC_function(5.3, spptomult="Arrowtooth", multiplier = 2, 
                 Pollock = 2e6, Arrowtooth = 2e5, Yellowfin = 2e5)#'                  
}
